+++ lib/main.dart
@@ -1,5 +1,9 @@
 
+
+
+import 'package:cura/individual/health1.dart';
 import 'package:cura/individual/home_page_individual.dart';
+import 'package:cura/individual/spend_time.dart';
 import 'package:cura/startup_screens/preview_page.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
