+++ pubspec.yaml
@@ -42,6 +42,9 @@ dependencies:
   searchfield: ^0.5.3
   smooth_page_indicator: ^1.0.0+2
   intl: ^0.17.0
+  dependencies:
+  dropdown_textfield: ^1.0.5
+  
 
 
 dev_dependencies:
